Massive code and docs cleanup

+++ kbox2d/src/commonMain/kotlin/org/jbox2d/common/Vec3.kt
@@ -29,12 +29,9 @@ import org.jbox2d.internal.*
  * @author Daniel Murphy
  */
 data class Vec3(
-
-    var x: Float = 0.toFloat(),
-
-    var y: Float = 0.toFloat(),
-
-    var z: Float = 0.toFloat()
+    var x: Float = 0f,
+    var y: Float = 0f,
+    var z: Float = 0f
 ) {
     constructor(copy: Vec3) : this(copy.x, copy.y, copy.z)
